[VariantBundle] add js to variant bundle
