+++ src/CoreShop/Bundle/VariantBundle/Resources/config/pimcore/config.yml
@@ -1,2 +1,3 @@
 imports:
+    - { resource: admin.yml }
     - { resource: messenger.yml }
